backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
motor==3.3.2
web3==6.15.1
python-dotenv==1.0.1
scikit-learn==1.4.0
bcrypt==4.1.2
PyJWT==2.8.0
pydantic[email]==2.6.0
websockets==12.0
redis==5.0.1
groq==0.4.2pymongo==4.6.1
faker==22.6.0
